+++ api/src/main/java/org/allaymc/api/blockentity/interfaces/BlockEntityChest.java
@@ -1,11 +1,11 @@
 package org.allaymc.api.blockentity.interfaces;
 
 import org.allaymc.api.blockentity.BlockEntity;
-import org.allaymc.api.blockentity.component.BlockEntityContainerHolderComponent;
+import org.allaymc.api.blockentity.component.BlockEntityChestContainerHolderComponent;
 import org.allaymc.api.blockentity.component.BlockEntityPairableComponent;
 
 /**
  * @author daoge_cmd
  */
-public interface BlockEntityChest extends BlockEntity, BlockEntityContainerHolderComponent, BlockEntityPairableComponent {
+public interface BlockEntityChest extends BlockEntity, BlockEntityChestContainerHolderComponent, BlockEntityPairableComponent {
 }