+++ api/src/main/java/org/allaymc/api/entity/component/EntityItemBaseComponent.java
@@ -9,14 +9,14 @@ public interface EntityItemBaseComponent extends EntityPickableBaseComponent {
     /**
      * Get the item stack of this entity item.
      *
-     * @return the item stack, can be null
+     * @return the item stack, can be {@code null}
      */
     ItemStack getItemStack();
 
     /**
      * Set the item stack of this entity item.
      *
-     * @param itemStack the item stack, can be null
+     * @param itemStack the item stack, can be {@code null}
      */
     void setItemStack(ItemStack itemStack);
 }